feat: Create my Encrypter and Decrypter

+++ decrypter.py
@@ -1,22 +1,26 @@
 import os
 import pyaes
 
-## abrir o arquivo criptografado
-file_name = "teste.txt.ransomwaretroll"
-file = open(file_name, "rb")
-file_data = file.read()
-file.close()
-
-## chave para descriptografia
-key = b"testeransomwares"
+# chave para descriptografia (deve ser igual à usada na criptografia)
+key = b"victorreis123456"
 aes = pyaes.AESModeOfOperationCTR(key)
-decrypt_data = aes.decrypt(file_data)
 
-## remover o arquivo criptografado
-os.remove(file_name)
+# descriptografar todos os arquivos .troll na pasta
+for file_name in os.listdir():
+    if file_name.endswith(".troll"):
+        # abrir o arquivo criptografado
+        with open(file_name, "rb") as file:
+            file_data = file.read()
 
-## criar o arquivo descriptografado
-new_file = "teste.txt"
-new_file = open(f'{new_file}', "wb")
-new_file.write(decrypt_data)
-new_file.close()
+        # descriptografar o arquivo
+        decrypt_data = aes.decrypt(file_data)
+
+        # remover o arquivo criptografado
+        os.remove(file_name)
+
+        # restaurar o arquivo original
+        original_file_name = file_name.replace(".troll", "")
+        with open(original_file_name, "wb") as new_file:
+            new_file.write(decrypt_data)
+
+print("Arquivos descriptografados com sucesso!")
